Bug 1827192 - Move main menu items related UI tests to a new class
