+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/SmokeTest.kt
@@ -113,120 +113,6 @@ class SmokeTest {
         }
     }
 
-    // Verifies the list of items in a tab's 3 dot menu
-    @Test
-    fun verifyPageMainMenuItemsTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            waitForPageToLoad()
-        }.openThreeDotMenu {
-            verifyPageThreeDotMainMenuItems(isRequestDesktopSiteEnabled = false)
-        }
-    }
-
-    // Could be removed when more smoke tests from the History category are added
-    // Verifies the History menu opens from a tab's 3 dot menu
-    @Test
-    fun openMainMenuHistoryItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openHistory {
-            verifyHistoryListExists()
-        }
-    }
-
-    // Could be removed when more smoke tests from the Bookmarks category are added
-    // Verifies the Bookmarks menu opens from a tab's 3 dot menu
-    @Test
-    fun openMainMenuBookmarksItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openBookmarks {
-            verifyBookmarksMenuView()
-        }
-    }
-
-    // Verifies the Add-ons menu opens from a tab's 3 dot menu
-    @Test
-    fun openMainMenuAddonsTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openAddonsManagerMenu {
-            registerAndCleanupIdlingResources(
-                RecyclerViewIdlingResource(activityTestRule.activity.findViewById(R.id.add_ons_list), 1),
-            ) {
-                verifyAddonsItems()
-            }
-        }
-    }
-
-    // Verifies the Synced tabs menu or Sync Sign In menu opens from a tab's 3 dot menu.
-    // The test is assuming we are NOT signed in.
-    @Test
-    fun openMainMenuSyncItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-            mDevice.waitForIdle()
-        }.openThreeDotMenu {
-        }.openSyncSignIn {
-            verifyTurnOnSyncMenu()
-        }
-    }
-
-    // Test running on beta/release builds in CI:
-    // caution when making changes to it, so they don't block the builds
-    // Verifies the Settings menu opens from a tab's 3 dot menu
-    @Test
-    fun openMainMenuSettingsItemTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openSettings {
-            verifySettingsView()
-        }
-    }
-
-    // Verifies the Find in page option in a tab's 3 dot menu
-    @Test
-    fun openMainMenuFindInPageTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openFindInPage {
-            verifyFindInPageSearchBarItems()
-        }
-    }
-
-    // Verifies the Add to collection option in a tab's 3 dot menu
-    @Test
-    fun openMainMenuAddToCollectionTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.openSaveToCollection {
-            verifyCollectionNameTextField()
-        }
-    }
-
     // Verifies the Bookmark button in a tab's 3 dot menu
     @Test
     fun mainMenuBookmarkButtonTest() {
@@ -252,50 +138,6 @@ class SmokeTest {
         }.openThreeDotMenu {
         }.clickOpenInApp {
             assertYoutubeAppOpens()
-        }
-    }
-
-    // Verifies the Desktop site toggle in a tab's 3 dot menu
-    @Test
-    fun mainMenuDesktopSiteTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.switchDesktopSiteMode {
-        }.openThreeDotMenu {
-            verifyDesktopSiteModeEnabled(true)
-        }
-    }
-
-    // Verifies the Share button in a tab's 3 dot menu
-    @Test
-    fun mainMenuShareButtonTest() {
-        val defaultWebPage = TestAssetHelper.getGenericAsset(mockWebServer, 1)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(defaultWebPage.url) {
-        }.openThreeDotMenu {
-        }.clickShareButton {
-            verifyShareTabLayout()
-            verifySendToDeviceTitle()
-            verifyShareALinkTitle()
-        }
-    }
-
-    // Verifies the refresh button in a tab's 3 dot menu
-    @Test
-    fun mainMenuRefreshButtonTest() {
-        val refreshWebPage = TestAssetHelper.getRefreshAsset(mockWebServer)
-
-        navigationToolbar {
-        }.enterURLAndEnterToBrowser(refreshWebPage.url) {
-            mDevice.waitForIdle()
-        }.openThreeDotMenu {
-            verifyThreeDotMenuExists()
-        }.refreshPage {
-            verifyPageContent("REFRESHED")
         }
     }
