+++ fenix/automation/taskcluster/androidTest/flank-arm-start-test.yml
@@ -22,7 +22,7 @@ gcloud:
   test-targets:
    - class org.mozilla.fenix.ui.NavigationToolbarTest#visitURLTest
    - class org.mozilla.fenix.ui.HistoryTest#visitedUrlHistoryTest
-   - class org.mozilla.fenix.ui.SmokeTest#openMainMenuSettingsItemTest
+   - class org.mozilla.fenix.ui.MainMenuTest#openMainMenuSettingsItemTest
    - class org.mozilla.fenix.ui.SettingsSearchTest#toggleSearchSuggestionsTest
    - class org.mozilla.fenix.ui.CollectionTest#deleteCollectionTest
    - class org.mozilla.fenix.ui.SmokeTest#noHistoryInPrivateBrowsingTest
